# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "PayableImportationResponse": {
        "description": "Importação assíncrona de contas a pagar.",
        "properties": {
          "createdAt": {
            "description": "Data e hora de criação.",
            "example": "2026-08-01T10:15:30",
            "type": "string"
          },
          "errorMessage": {
            "description": "Erro global quando a importação estiver em FAILED.",
            "type": "string"
          },
          "id": {
            "description": "Identificador da importação.",
            "example": 1,
            "format": "int64",
            "type": "integer"
          },
          "status": {
            "description": "Estado do processamento assíncrono.",
            "enum": [
              "PENDING",
              "PROCESSING",
              "COMPLETED",
              "COMPLETED_WITH_ERRORS",
              "FAILED"
            ],
            "example": "PENDING",
            "type": "string"
          },
          "updatedAt": {
            "description": "Data e hora da última atualização.",
            "example": "2026-08-01T10:15:35",
            "type": "string"
          }
        },
        "type": "object"
      },
      "PayablePageResponse": {
        "description": "Página de contas a pagar.",
        "properties": {
          "content": {
            "items": {
              "$ref": "#/components/schemas/PayableResponse"
            },
            "type": "array"
          },
          "empty": {
            "type": "boolean"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "number": {
            "format": "int32",
            "type": "integer"
          },
          "numberOfElements": {
            "format": "int32",
            "type": "integer"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableResponse"
          },
          "size": {
            "format": "int32",
            "type": "integer"
          },
          "sort": {
            "$ref": "#/components/schemas/SortResponse"
          },
          "totalElements": {
            "format": "int64",
            "type": "integer"
          },
          "totalPages": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "ProblemDetail": {
        "description": "Erro HTTP no formato RFC 9457.",
        "properties": {
          "detail": {
            "description": "Explicação específica do erro.",
            "example": "Dados inválidos",
            "type": "string"
          },
          "instance": {
            "description": "URI da requisição que originou o erro.",
            "type": "string"
          },
          "status": {
            "description": "Código HTTP.",
            "example": 400,
            "format": "int32",
            "type": "integer"
          },
          "title": {
            "description": "Título correspondente ao status HTTP.",
            "example": "Bad Request",
            "type": "string"
          },
          "type": {
            "description": "Identificador do tipo do problema.",
            "example": "about:blank",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "basicAuth": {
        "description": "Credenciais configuradas por APP_SECURITY_USERNAME e APP_SECURITY_PASSWORD.",
        "scheme": "basic",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API para gestão de fornecedores, contas a pagar, importações assíncronas em CSV\ne relatórios de pagamentos. Todos os recursos exigem autenticação HTTP Basic.\n",
    "title": "Payable Management API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/import/payable/{id}": {
      "get": {
        "description": "Retorna uma importação pelo identificador para acompanhamento assíncrono.",
        "operationId": "findPayableImportationById",
        "parameters": [
          {
            "description": "Identificador da importação.",
            "example": 1,
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayableImportationResponse"
                }
              }
            },
            "description": "Importação encontrada."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Credenciais de autenticação ausentes ou inválidas",
                  "status": 401,
                  "title": "Unauthorized",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Credenciais ausentes ou inválidas."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Recurso não encontrado",
                  "status": 404,
                  "title": "Not Found",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Importação não encontrada."
          }
        },
        "summary": "Consultar importação",
        "tags": [
          "Importações"
        ]
      }
    },
...
    "/payable": {
      "get": {
        "description": "Retorna contas paginadas. As datas filtram o vencimento e devem ser informadas\njuntas; a data inicial não pode ser posterior à final.\n",
        "operationId": "listPayables",
        "parameters": [
          {
            "description": "Número da página, iniciado em zero.",
            "in": "query",
            "name": "page",
            "schema": {
              "default": 0,
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "description": "Quantidade de registros por página.",
            "in": "query",
            "name": "size",
            "schema": {
              "default": 20,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Ordenação no formato propriedade,direção. Pode ser repetido.",
            "example": "expirationDate,asc",
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Trecho da descrição, sem distinção entre maiúsculas e minúsculas.",
            "example": "aluguel",
            "in": "query",
            "name": "description",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Início opcional do período de vencimento.",
            "example": "2026-08-01",
            "in": "query",
            "name": "startDate",
            "required": false,
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "description": "Fim opcional do período de vencimento.",
            "example": "2026-08-31",
            "in": "query",
            "name": "endDate",
            "required": false,
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "description": "Status por nome, sem distinção entre maiúsculas e minúsculas.",
            "example": "PAGO",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "PENDENTE",
                "PAGO",
                "CANCELADO"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayablePageResponse"
                }
              }
            },
            "description": "Página de contas a pagar."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Dados inválidos",
                  "status": 400,
                  "title": "Bad Request",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Filtro de status ou período inválido."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Credenciais de autenticação ausentes ou inválidas",
                  "status": 401,
                  "title": "Unauthorized",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Credenciais ausentes ou inválidas."
          }
        },
        "summary": "Listar contas a pagar",
        "tags": [
          "Contas a pagar"
        ]
      },
      "post": {
        "description": "Cadastra uma conta para um fornecedor existente. O status padrão é PENDENTE.\nUma conta PAGO exige paymentDate e outros status exigem essa data nula.\n",
        "operationId": "createPayable",
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "amount": 1500.0,
                "description": "Pagamento de aluguel",
                "expirationDate": "2026-08-10",
                "status": "PENDENTE",
                "supplierId": 1
              },
              "schema": {
                "$ref": "#/components/schemas/PayableRequest"
              }
            }
          },
          "description": "Dados da nova conta a pagar.",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayableResponse"
                }
              }
            },
            "description": "Conta a pagar criada."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Dados inválidos",
                  "status": 400,
                  "title": "Bad Request",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Dados violam uma regra da conta a pagar."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Credenciais de autenticação ausentes ou inválidas",
                  "status": 401,
                  "title": "Unauthorized",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Credenciais ausentes ou inválidas."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Recurso não encontrado",
                  "status": 404,
                  "title": "Not Found",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Fornecedor informado não existe."
          }
        },
        "summary": "Criar conta a pagar",
        "tags": [
          "Contas a pagar"
        ]
      }
    },
...
    "/payable/{id}/status": {
      "patch": {
        "description": "Atualiza somente o status. Ao passar para PAGO sem paymentDate, utiliza a data atual\nou preserva a data anterior se a conta já estava paga. Outros status removem a data.\n",
        "operationId": "updatePayableStatus",
        "parameters": [
          {
            "description": "Identificador da conta a pagar.",
            "example": 1,
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "paymentDate": "2026-08-10",
                "status": 1
              },
              "schema": {
                "$ref": "#/components/schemas/UpdatePayableStatusRequest"
              }
            }
          },
          "description": "Código do novo status e data de pagamento opcional.",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayableResponse"
                }
              }
            },
            "description": "Status atualizado ou conta retornada sem alteração."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Dados inválidos",
                  "status": 400,
                  "title": "Bad Request",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Código de status ausente, inexistente ou combinação inválida."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Credenciais de autenticação ausentes ou inválidas",
                  "status": 401,
                  "title": "Unauthorized",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Credenciais ausentes ou inválidas."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Recurso não encontrado",
                  "status": 404,
                  "title": "Not Found",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conta a pagar não encontrada."
          }
        },
        "summary": "Alterar status da conta",
        "tags": [
          "Contas a pagar"
        ]
      }
    },
...
  },
  "security": [
    {
      "basicAuth": []
    }
  ],
  "servers": [
    {
      "description": "Servidor da aplicação",
      "url": "/management/rest"
    }
  ],
...
}